Add files via upload
The titanic and titanic2 data frames describe the survival status of individual passengers
on the Titanic. The titanic data frame does not contain information from the crew, but it
does contain actual ages of half of the passengers. The principal source for data about
Titanic passengers is the Encyclopedia Titanica. The datasets used here were begun by
a variety of researchers. One of the original sources is Eaton & Haas (1994) Titanic:
Triumph and Tragedy, Patrick Stephens Ltd, which includes a passenger list created by
many researchers and edited by Michael A. Findlay.
In this Data set, I will explore and analysis data and use visualization using matplotlib and seaborn and answer many questions about it